Reorganize to reflect function used in multiple analyses
