+++ vdmlab/__init__.py
@@ -1,4 +1,4 @@
-from .co_occurrence import spike_counts, get_tetrode_mask, find_multi_in_epochs, compute_cooccur
+from .co_occurrence import spike_counts, get_tetrode_mask, compute_cooccur
 from .decoding import bayesian_prob, decode_location, remove_teleports
 from .lfp_filtering import detect_swr_hilbert
 from .maze_breakdown import expand_line, save_spike_position
@@ -10,6 +10,7 @@ from .objects import (AnalogSignal,
 from .place_fields import consecutive, find_fields, get_single_field, get_heatmaps
 from .tuning_curves import binned_position, tuning_curve, tuning_curve_2d
 from .utils import (find_nearest_idx,
+                    find_multi_in_epochs,
                     get_sort_idx,
                     add_scalebar,
                     get_counts,
