+++ vdmlab/co_occurrence.py
@@ -50,28 +50,6 @@ def get_tetrode_mask(spikes):
     return tetrode_mask
 
 
-def find_multi_in_epochs(spikes, epochs, min_involved):
-    multi_starts = []
-    multi_stops = []
-
-    n_neurons = len(spikes)
-    for start, stop in zip(epochs.starts, epochs.stops):
-        involved = 0
-        for neuron in range(n_neurons):
-            if ((start <= spikes[neuron].time) & (spikes[neuron].time <= stop)).sum() > 1:
-                involved += 1
-        if involved > min_involved:
-            multi_starts.append(start)
-            multi_stops.append(stop)
-
-    multi_starts = np.array(multi_starts)
-    multi_stops = np.array(multi_stops)
-
-    multi_epochs = vdm.Epoch(np.array([multi_starts, multi_stops]))
-
-    return multi_epochs
-
-
 def compute_cooccur(count_matrix, tetrode_mask, num_shuffles=10000):
     """Computes the probabilities for co-occurrence
